Restructure files into nicer pattern
