+++ handlers.go
@@ -2,11 +2,6 @@ package main
 
 type Handler func(*Client, interface{})
 
-type Message struct {
-	Name string      `json:"name"`
-	Data interface{} `json:"data"`
-}
-
 func addFeed(client *Client, data interface{}) {
 
 }
