+++ message.go
@@ -1,0 +1,6 @@
+package main
+
+type Message struct {
+	Name string      `json:"name"`
+	Data interface{} `json:"data"`
+}
